Added responsive layout and margin adjustment
This update brings improvements to the layout of the SearchForm component. Specifically, a responsive design has been added to handle screens with a width of less than 600px. In this case, the layout switches from a row to a column, and input fields and buttons stretch to occupy 80% of the screen width. Moreover, the margin around the form has been increased from 20px to 40px for better spacing.