Added transition to Hamburger Menu as well as supporting libraries
In this commit, the Navbar component was updated to display its menu items in a Hamburger Menu on smaller screens for better mobile user experience. To facilitate this, necessary libraries were added to support the functionality of the Hamburger Menu. This enhances the application's responsiveness and usability on mobile devices.