Added AboutUs component
This commit includes the creation of the new AboutUs component, which provides a description of the project to users. This component has been included in App.js, which is the main entry point of the application.

Moreover, a link to the AboutUs component has been added to the navigation bar, Navbar.js, to make it easily accessible from anywhere in the application.

These changes improve the navigability of the application and provide users with easy access to information about the project.