Added support for sending credentials with Axios requests
This commit includes a modification in the App.js file to set the withCredentials property to true for all Axios requests. This is done by adding the line:

axios.defaults.withCredentials = true;

This setting ensures that Axios sends cookies and other credentials such as the OAuth token along with the request. It is crucial when making authenticated requests to a server that requires these credentials, for example, when trying to connect to a server with CORS enabled. This change allows for proper communication with backend services requiring authentication.