Updated MovieList.js, CSS, for movie data grouping and presentation
Updated MovieList.js to sort and group movies after receiving data from the API, and improved the related CSS and JSX files for a better data presentation.

We first sort the movies alphabetically by cinema name, then by movie title. Next, we group the movies by cinema. The result is an object where the keys are cinema names, and the values are arrays of movies.

In rendering, we create a section for each cinema with the cinema's name and a list of its movies. The CSS and JSX files (MovieList.css, MovieItem.css, MovieList.js) have been updated for better presentation of movie data. The movie list's appearance has been improved to be more readable and appealing to users.

Readability on mobile devices is ensured by using Responsive Web Design techniques.